// src/generated/zod.gen.ts
export const zPlatform = z.enum([
  'AMAZON',
  'EBAY',
  'ETSY',
  'FNAC',
  'WALMART',
  'SHOPEE',
  'LAZADA',
  'MERCADOLIBRE',
  'ALIBABA',
  'SHOPLINE',
  'PRESTASHOP',
  'SHOPIFY',
  'MAGENTO',
  'WOO_COMMERCE',
  'SQUARESPACE',
  'WIX',
  'SHOPER',
  'SHOPLAZZA',
  'BIG_COMMERCE',
  'SALLA',
  'ZID',
  'COMMERCELAYER',
  'WEBFLOW',
  'GUMROAD',
  'ECWID',
  'PAYPAL',
  'SQUARE',
  'STRIPE',
  'CLOVER',
  'MOLLIE',
  'PAYNL',
  'AUTHORIZENET',
  'RECHARGE',
  'CHARGIFY',
  'CHARGEBEE',
  'RECURLY',
  'KASHFLOW',
  'SAGE_INTACCT',
  'SAGE_INTACCT_V2',
  'DYNAMICS365',
  'ZOHOBOOKS',
  'EXACTONLINE',
  'FREEAGENT',
  'QUICKBOOKS',
  'QUICKBOOKS_DESKTOP',
  'FRESHBOOKS',
  'XERO',
  'SAGE_BUSINESS_CLOUD',
  'SAGE_50',
  'NETSUITE',
  'WAVE',
  'DUALENTRY',
  'MONEYBIRD',
  'SAGE200CLOUD',
  'MX',
  'PLAID',
  'TELLER',
  'ODOO',
  'TALLY',
  'GOOGLE',
  'FACEBOOK',
  'TIKTOK',
  'INTUIT_BANK_FEEDS',
  'SAGE',
  'SAGE_BANK_FEEDS',
  'QUICKBOOKS_HRIS',
  'BAMBOO_HR',
  'GUSTO',
])
// ...
export const zListConnectionsResponse = z.object({
  connections: z.array(
    z.object({
      id: z.string(),
      access_token: z.string(),
      link_url: z.string(),
      platform: z.string(),
      status: z.enum([
        'active',
        'unhealthy',
        'not_ready',
        'needs_reauth',
        'disabled',
      ]),
      name: z.string().nullable(),
      created_at: z.string(),
      updated_at: z.string(),
    }),
  ),
})

export const zUpdateConnectionBody = z.object({
  name: z.string().max(30),
})

export const zConnectionResponse = z.object({
  connection: z.object({
    id: z.string(),
    access_token: z.string(),
    link_url: z.string(),
    name: z.string().nullable(),
  }),
})

export const zDeleteConnectionPathParams = z.object({
  id: z.string(),
})

export const zDeleteConnectionResponse = z.object({
  success: z.boolean(),
})

export const zGetAccessTokenConnectionResponse = z.object({
  connection: z.object({
    id: z.string(),
    store_unique_id: z.string(),
    estimated_completed_at: z.string().nullish(),
    last_sync_completed_at: z.string().nullish(),
    last_sync_started_at: z.string().nullish(),
    oldest_order_date: z.string().nullish(),
    newest_order_date: z.string().nullish(),
    disabled: z.boolean(),
    link_url: z.string(),
    needs_update: z.boolean(),
    platform: zPlatform,
    initial_orders_synced_count: z.number().optional(),
    store_domain: z.string().nullable(),
    store_name: z.string().nullable(),
    unavailable_objects: z
      .array(
        z.enum([
          'account',
          'accounting_customer',
          'accounting_customer_refund',
          'accounting_transaction',
          'account_transaction',
          'attachment',
          'balance_sheet',
          'bank_deposit',
          'bank_feed_account',
          'bank_feed_transaction',
          'bank_transfer',
          'bill',
          'bill_credit_memo',
          'bill_payment',
          'cash_flow',
          'categories',
          'class',
          'commerce_customer',
          'company_info',
          'currencies',
          'customer_group',
          'department',
          'expense',
          'income_statement',
          'invoice',
          'invoice_credit_memo',
          'invoice_payment',
          'items',
          'journal_entry',
          'location',
          'order',
          'payment_method',
          'payment_term',
          'payout',
          'product',
          'project',
          'purchase_order',
          'sales_order',
          'sales_receipt',
          'store',
          'subscription',
          'subsidiary',
          'task',
          'tax_agency',
          'tax_area',
          'tax_rates',
          'transaction',
          'updated_transaction',
          'vendor',
        ]),
      )
      .nullable(),
    created_at: z.string().optional(),
  }),
})
// ...
export const zExchangeTokenResponse = z.object({
  connection_id: z.string(),
  access_token: z.string(),
  is_ready: z.boolean(),
  platform: z.enum([
    'AMAZON',
    'EBAY',
    'ETSY',
    'FNAC',
    'WALMART',
    'SHOPEE',
    'LAZADA',
    'MERCADOLIBRE',
    'ALIBABA',
    'SHOPLINE',
    'PRESTASHOP',
    'SHOPIFY',
    'MAGENTO',
    'WOO_COMMERCE',
    'SQUARESPACE',
    'WIX',
    'SHOPER',
    'SHOPLAZZA',
    'BIG_COMMERCE',
    'SALLA',
    'ZID',
    'COMMERCELAYER',
    'WEBFLOW',
    'GUMROAD',
    'ECWID',
    'PAYPAL',
    'SQUARE',
    'STRIPE',
    'CLOVER',
    'MOLLIE',
    'PAYNL',
    'AUTHORIZENET',
    'RECHARGE',
    'CHARGIFY',
    'CHARGEBEE',
    'RECURLY',
    'KASHFLOW',
    'SAGE_INTACCT',
    'SAGE_INTACCT_V2',
    'DYNAMICS365',
    'ZOHOBOOKS',
    'EXACTONLINE',
    'FREEAGENT',
    'QUICKBOOKS',
    'QUICKBOOKS_DESKTOP',
    'FRESHBOOKS',
    'XERO',
    'SAGE_BUSINESS_CLOUD',
    'SAGE_50',
    'NETSUITE',
    'WAVE',
    'DUALENTRY',
    'MONEYBIRD',
    'SAGE200CLOUD',
    'MX',
    'PLAID',
    'TELLER',
    'ODOO',
    'TALLY',
    'GOOGLE',
    'FACEBOOK',
    'TIKTOK',
    'INTUIT_BANK_FEEDS',
    'SAGE',
    'SAGE_BANK_FEEDS',
    'QUICKBOOKS_HRIS',
    'BAMBOO_HR',
    'GUSTO',
  ]),
  name: z.string().nullable(),
  store_unique_name: z.string().optional(),
})